+++ compiler/GHC/StgToCmm/Prim.hs
@@ -1624,6 +1624,7 @@ emitPrimOp cfg primop =
   MkApUpd0_Op -> alwaysExternal
   NewBCOOp -> alwaysExternal
   UnpackClosureOp -> alwaysExternal
+  ListThreadsOp -> alwaysExternal
   ClosureSizeOp -> alwaysExternal
   WhereFromOp   -> alwaysExternal
   GetApStackValOp -> alwaysExternal